Refactor to have a class
