+++ test.js
@@ -76,11 +76,11 @@ function show (arr) {
 
 
 
-let N = 1e3
+let N = 1e6
 let ratio = window.innerWidth / window.innerHeight
 let range = [-10 * ratio, -10, 10 * ratio, 10]
 let colors = palettes[Math.floor(Math.random() * palettes.length)]
-let markers = [null, dist]
+let markers = [null]//, dist]
 let passes = markers.length
 
 let scatter = createScatter(regl)
